Abstract class, Interface implementation and Instantiation
It was created an abstract class to demonstrate how it can be used.  It
was also implemented the list interface to have a collection behavior.
Last but not least, it was created a set of constructors for Task class
with the idea of override default behavior and display a set of messages